refactor(hir): split Annex B.3.3.3 hoisting into its own module
`const_fold_fn.rs` crossed the 2000-line workspace file-size gate after the
B.3.3.3 hoist work. Move the self-contained hoisting machinery (synthesis
helpers, identifier renamer, `GlobalEvalHoist`, `apply_global_eval_hoist`, and
its tests) into `lower/global_eval_hoist.rs`; `const_fold_fn` calls the now
`pub(super)` entry point. Pure code move — no behavior change (annexB eval-code
sweep stays 459/469).
